third_party/blink/web_tests/external/wpt/css/css-fonts/variations/at-font-face-descriptors-expected.txt

This is a testharness.js-based test.
Found 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] font-weight(valid): Simple calc value: calc(100.5)
  assert_equals: Unexpected resulting value. expected "100.5" but got "calc(100.5)"
[FAIL] font-weight(valid): Valid calc expression: calc(100.5*3 + 50.5)
  assert_equals: Unexpected resulting value. expected "352" but got "calc(352)"
[FAIL] font-weight(valid): Simple range with equal upper and lower bounds: 500 500
  assert_equals: Unexpected resulting value. expected "500" but got "500 500"
[FAIL] font-weight(valid): Lower bound calc(): calc(100 + 100) 400
  assert_equals: Unexpected resulting value. expected "200 400" but got "calc(200) 400"
[FAIL] font-weight(valid): Upper bound calc(): 200 calc(200 + 200)
  assert_equals: Unexpected resulting value. expected "200 400" but got "200 calc(400)"
[FAIL] font-weight(valid): Both bounds are calc(): calc(100 + 100) calc(200 + 200)
  assert_equals: Unexpected resulting value. expected "200 400" but got "calc(200) calc(400)"
[FAIL] font-stretch(valid): Simple calc value: calc(200.5%)
  assert_equals: Unexpected resulting value. expected "200.5%" but got "calc(200.5%)"
[FAIL] font-stretch(valid): Valid calc expression: calc(50%*2 - 20%)
  assert_equals: Unexpected resulting value. expected "80%" but got "calc(80%)"
[FAIL] font-stretch(valid): Simple range with equal upper and lower bounds: 100% 100%
  assert_equals: Unexpected resulting value. expected "100%" but got "100% 100%"
[FAIL] font-stretch(valid): Lower bound calc(): calc(10% + 10%) 30%
  assert_equals: Unexpected resulting value. expected "20% 30%" but got "calc(20%) 30%"
[FAIL] font-stretch(valid): Upper bound calc(): 10% calc(10% + 10%)
  assert_equals: Unexpected resulting value. expected "10% 20%" but got "10% calc(20%)"
[FAIL] font-stretch(valid): Both bounds are calc(): calc(10% + 10%) calc(20% + 20%)
  assert_equals: Unexpected resulting value. expected "20% 40%" but got "calc(20%) calc(40%)"
[FAIL] font-style(valid): 'oblique' followed by calc with out of range value (should be clamped): oblique calc(91deg)
  assert_equals: Unexpected resulting value. expected "oblique calc(91deg)" but got "oblique 90deg"
[FAIL] font-style(valid): 'oblique' followed by calc with out of range value (should be clamped): oblique calc(-91deg)
  assert_equals: Unexpected resulting value. expected "oblique calc(-91deg)" but got "oblique -90deg"
[FAIL] font-style(valid): 'oblique' followed by angle in radians: oblique 0rad
  assert_equals: Unexpected resulting value. expected "oblique 0deg" but got "oblique 0rad"
[FAIL] font-style(valid): Simple range with equal upper and lower bounds: oblique 10deg 10deg
  assert_equals: Unexpected resulting value. expected "oblique 10deg" but got "oblique 10deg 10deg"
Harness: the test ran to completion.
